Cleanup the exports of sc-network
